refactor: move builtin names to separate file
